+++ presentation/src/main/java/com/cheocharm/presentation/ui/login/SignInActivity.kt
@@ -1,71 +1,24 @@
 package com.cheocharm.presentation.ui.login
 
-import android.content.Intent
 import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import androidx.activity.result.ActivityResultLauncher
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.activity.viewModels
+import androidx.navigation.fragment.NavHostFragment
 import com.cheocharm.base.BaseActivity
-import com.cheocharm.presentation.BuildConfig
 import com.cheocharm.presentation.R
 import com.cheocharm.presentation.databinding.ActivitySignInBinding
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import com.google.android.gms.auth.api.signin.GoogleSignInClient
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
-import com.google.android.gms.common.api.ApiException
-import com.google.android.gms.tasks.Task
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 class SignInActivity : BaseActivity<ActivitySignInBinding>(R.layout.activity_sign_in) {
 
-    private val loginViewModel by viewModels<LoginViewModel>()
-
-    private lateinit var googleSignInClient: GoogleSignInClient
-
-    private lateinit var intentResult: ActivityResultLauncher<Intent>
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-            .requestIdToken(BuildConfig.GOOGLE_LOGIN_WEB_CLIENT_ID)
-            .requestEmail()
-            .build()
-
-        googleSignInClient = GoogleSignIn.getClient(this, gso)
-
-        intentResult = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
-            if (it.resultCode == RESULT_OK) {
-                val task = GoogleSignIn.getSignedInAccountFromIntent(it.data)
-                handleSignInResult(task)
-            }
-        }
-        initButton()
+        setUp()
     }
 
-    override fun onStart() {
-        super.onStart()
-
-        val account = GoogleSignIn.getLastSignedInAccount(this)
-    }
-
-    private fun initButton() {
-        binding.btnSignInGoogleSignUp.setOnClickListener {
-            val intent = googleSignInClient.signInIntent
-            intentResult.launch(intent)
-        }
-    }
-
-    private fun handleSignInResult(completedTask: Task<GoogleSignInAccount>) {
-        try {
-            val account = completedTask.getResult(ApiException::class.java)
-
-        } catch (e: ApiException) {
-            Log.w("handleSignInResult", "${e.message}")
-        }
+    private fun setUp() {
+        val host: NavHostFragment =
+            (supportFragmentManager.findFragmentById(R.id.nav_host_fragment_login)) as NavHostFragment
+        val navController = host.navController
     }
 }
