+++ presentation/src/main/java/com/cheocharm/presentation/ui/login/SignInFragment.kt
@@ -1,0 +1,77 @@
+package com.cheocharm.presentation.ui.login
+
+import android.content.Intent
+import android.os.Bundle
+import android.util.Log
+import android.view.View
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.viewModels
+import com.cheocharm.base.BaseFragment
+import com.cheocharm.presentation.BuildConfig
+import com.cheocharm.presentation.R
+import com.cheocharm.presentation.databinding.FragmentSignInBinding
+import com.google.android.gms.auth.api.signin.GoogleSignIn
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount
+import com.google.android.gms.auth.api.signin.GoogleSignInClient
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions
+import com.google.android.gms.common.api.ApiException
+import com.google.android.gms.tasks.Task
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class SignInFragment : BaseFragment<FragmentSignInBinding>(R.layout.fragment_sign_in) {
+
+    private val loginViewModel by viewModels<LoginViewModel>()
+
+    private lateinit var googleSignInClient: GoogleSignInClient
+
+    private lateinit var intentResult: ActivityResultLauncher<Intent>
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        val gso = GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+            .requestIdToken(BuildConfig.GOOGLE_LOGIN_WEB_CLIENT_ID)
+            .requestEmail()
+            .build()
+
+        googleSignInClient = GoogleSignIn.getClient(requireActivity(), gso)
+
+        intentResult = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
+            if (it.resultCode == AppCompatActivity.RESULT_OK) {
+                val task = GoogleSignIn.getSignedInAccountFromIntent(it.data)
+                handleSignInResult(task)
+            }
+        }
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        initButton()
+    }
+
+    override fun onStart() {
+        super.onStart()
+
+        val account = GoogleSignIn.getLastSignedInAccount(requireActivity())
+    }
+
+    private fun initButton() {
+        binding.btnSignInGoogleSignUp.setOnClickListener {
+            val intent = googleSignInClient.signInIntent
+            intentResult.launch(intent)
+        }
+    }
+
+    private fun handleSignInResult(completedTask: Task<GoogleSignInAccount>) {
+        try {
+            val account = completedTask.getResult(ApiException::class.java)
+
+        } catch (e: ApiException) {
+            Log.w("handleSignInResult", "${e.message}")
+        }
+    }
+}
